+++ CHANGELOG.md
@@ -6,12 +6,35 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.1.0] - Aug 07, 2022
+
+### Added
+
+* Added support for *Basic Authentication* using the `--username` and `--password` options.
+* Added support for custom *HTTP headers* using the `-H` option.
+
+### Changed
+
+* *webshell-cli* now uses patterns to detect webshell output within of server responses.
+  This is useful, when you have to embed the webshell into other content (e.g. a *JPEG* file).
+
+### Checksums (SHA256):
+
+* `webshell.php`: `812cbbf7fd27976ab0576ead9f565c54eb2f41500c7b3da9bde06f0f7c6f89e6`
+* `webshell.jsp`: `1c06b43aa06decd1d2a5e1be4a4aeb6c4db325f42018dcd920eeb686c9108586`
+* `webshell.aspx`: `6294b0be29209434152ac7c16980229e5d44a0d70ee514940d611f22a1e2441a`
+* `webshell-cli.py`: `aa3f7138537e1680d2c623f01faa7bbc7c5b7e1a25a338713c30ae02b3dea021`
+
+
 ## [1.0.0] - Apr 29, 2022
 
 ### Initial Release
 
-* Checksums (SHA256):
-    - `webshell.php`: `df8568952cb93a5ffb4becef0f31c98988a434eef2819b372aad78396e0c663f`
-    - `webshell.jsp`: `9e909081e9e74de4a798eb4be9118c037d34ad7f91e5b4a17dc5e9f736b5c156`
-    - `webshell.aspx`: `95f7caca59183eaa899b57946c9e5e9dbc1a96c077dce2b416228fe764ecee88`
-    - `webshell-cli.py`: `e8a879bf19acccc3ee490bd19527ed9459b749d8108c722ebac7e52fe6dcd953`
+:)
+
+### Checksums (SHA256):
+
+* `webshell.php`: `df8568952cb93a5ffb4becef0f31c98988a434eef2819b372aad78396e0c663f`
+* `webshell.jsp`: `9e909081e9e74de4a798eb4be9118c037d34ad7f91e5b4a17dc5e9f736b5c156`
+* `webshell.aspx`: `95f7caca59183eaa899b57946c9e5e9dbc1a96c077dce2b416228fe764ecee88`
+* `webshell-cli.py`: `e8a879bf19acccc3ee490bd19527ed9459b749d8108c722ebac7e52fe6dcd953`